+++ AdventOfCode2019/AdventOfCode2019.Puzzles.Intcode/Instructions/InstructionBase.cs
@@ -19,14 +19,14 @@ namespace AdventOfCode2019.Puzzles.Intcode.Instructions
             ProgramContext = context;
         }
 
-        protected int GetParameterValue(int parameterPosition)
+        protected long GetParameterValue(int parameterPosition)
         {
             int digit = (int)Math.Pow(10, parameterPosition) * 100;
             var instructionCode = ProgramContext.Memory[ProgramContext.InstructionPointer];
 
             if ((instructionCode % digit * 10 / digit) == 0) // absolute position mode
             {
-                return ProgramContext.Memory[0 + ProgramContext.Memory[ProgramContext.InstructionPointer + parameterPosition]];
+                return ProgramContext.Memory[0 + (int)ProgramContext.Memory[ProgramContext.InstructionPointer + parameterPosition]];
             }
             else if ((instructionCode % digit * 10 / digit) == 1) // immediate mode (value)
             {
@@ -34,7 +34,7 @@ namespace AdventOfCode2019.Puzzles.Intcode.Instructions
             }
             else if ((instructionCode % digit * 10 / digit) == 2) // relative position mode 
             {
-                return ProgramContext.Memory[ProgramContext.RelativeBase + ProgramContext.Memory[ProgramContext.InstructionPointer + parameterPosition]];
+                return ProgramContext.Memory[ProgramContext.RelativeBase + (int)ProgramContext.Memory[ProgramContext.InstructionPointer + parameterPosition]];
             }
             else
             {
