+++ AdventOfCode2019/AdventOfCode2019.Puzzles/Extensions/StringExtensions.cs
@@ -9,22 +9,22 @@ namespace AdventOfCode2019.Puzzles.Tests.Extensions
 {
     public static class ChannelExtensions
     {
-        public static async Task<Channel<int>> ToChannelAsync(this string s)
+        public static async Task<Channel<long>> ToChannelAsync(this string s)
         {
-            var c = Channel.CreateUnbounded<int>();
+            var c = Channel.CreateUnbounded<long>();
             using var stringReader = new StringReader(s);
             string line = null;
             while ((line = await stringReader.ReadLineAsync()) != null)
             {
-                await c.Writer.WriteAsync(int.Parse(line));
+                await c.Writer.WriteAsync(long.Parse(line));
             }
 
             return c;
         }
 
-        public static async Task<Channel<int>> ToChannelAsync(this int x)
+        public static async Task<Channel<long>> ToChannelAsync(this int x)
         {
-            var c = Channel.CreateUnbounded<int>();
+            var c = Channel.CreateUnbounded<long>();
             await c.Writer.WriteAsync(x);
             return c;
         }
